src/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-RESTful==0.3.10
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pymongo==4.6.3
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
Werkzeug==3.0.2
